feat: product to sqs
