+++ import-service/lib/import-service-stack.ts
@@ -7,6 +7,8 @@ import * as apigw from '@aws-cdk/aws-apigatewayv2-alpha';
 import { HttpLambdaIntegration } from '@aws-cdk/aws-apigatewayv2-integrations-alpha';
 import { EventType } from 'aws-cdk-lib/aws-s3';
 import { LambdaDestination } from 'aws-cdk-lib/aws-s3-notifications';
+import * as sqs from 'aws-cdk-lib/aws-sqs';
+import { NodejsFunction } from 'aws-cdk-lib/aws-lambda-nodejs';
 
 export const BUCKET_NAME = 'import-bucket-st';
 
@@ -46,6 +48,8 @@ export class ImportServiceStack extends cdk.Stack {
       ],
     });
 
+    const queue = sqs.Queue.fromQueueArn(this, 'ImportQueue', 'arn:aws:sqs:eu-west-1:147793823999:importQueue');
+
     const importProductsFileLambda = new lambda.Function(
       this,
       'importProductsFileLambdaSt',
@@ -58,28 +62,27 @@ export class ImportServiceStack extends cdk.Stack {
         environment: {
           BUCKET: BUCKET_NAME,
           REGION: 'eu-west-1',
+          SQS_URL: queue.queueUrl,
         },
       }
     );
 
-    const importFileParserLambda = new lambda.Function(
-      this,
-      'importFileParserLambdaSt',
-      {
-        functionName: 'import-file-parser',
-        runtime: lambda.Runtime.NODEJS_18_X,
-        handler: 'importFileParser.handler',
-        code: lambda.Code.fromAsset('lambda'),
-        role: iAmRole,
-        environment: {
-          BUCKET: BUCKET_NAME,
-          REGION: 'eu-west-1',
-        },
-      }
-    );
+    const importFileParserLambda = new NodejsFunction(this, "importFileParserLambdaSt", {
+      functionName: 'import-file-parser',
+      runtime: lambda.Runtime.NODEJS_18_X,
+      handler: 'handler',
+      entry: 'lambda/importFileParser.ts',
+      role: iAmRole,
+      environment: {
+        BUCKET: BUCKET_NAME,
+        REGION: 'eu-west-1',
+        SQS_URL: queue.queueUrl,
+      },
+    });
 
     importBucket.grantReadWrite(importProductsFileLambda);
     importBucket.grantReadWrite(importFileParserLambda);
+    queue.grantSendMessages(importFileParserLambda);
 
     const apiGateway = new apigw.HttpApi(this, 'importApiGateway', {
       corsPreflight: {
