+++ import-service/package.json
@@ -14,6 +14,7 @@
     "@types/jest": "^29.5.8",
     "@types/node": "20.9.0",
     "aws-cdk": "2.110.1",
+    "esbuild": "^0.19.9",
     "jest": "^29.7.0",
     "ts-jest": "^29.1.1",
     "ts-node": "^10.9.1",
@@ -29,6 +30,7 @@
     "aws-sdk": "^2.1511.0",
     "constructs": "^10.0.0",
     "csv-parser": "^3.0.0",
+    "fast-csv": "^4.3.6",
     "source-map-support": "^0.5.21"
   }
 }
